refactor(Utils): split up Utils\Utils into Utils\ArrayHelpers and Utils\Asset
Asset is responsible for getting the correct url and path (after revving)
